Fix VAD/Moonshine loading race condition
This commit moves all transcription logic into Transcriber and removes StreamTranscriber. This consolidates the initialization of the VAD and speech-to-text models into one class, allowing the load of both models to occur in a single function.
