+++ docs/api/classes/_index.md
@@ -2,10 +2,10 @@
 title: "Classes"
 ---
 - [IntentClassifier](/docs/api/classes/intentclassifier)
-- [KeywordSpotter](/docs/api/classes/keywordspotter)
 - [MediaElementTranscriber](/docs/api/classes/mediaelementtranscriber)
+- [KeywordSpotter](/docs/api/classes/keywordspotter)
 - [MicrophoneTranscriber](/docs/api/classes/microphonetranscriber)
 - [MoonshineModel](/docs/api/classes/moonshinemodel)
 - [MoonshineSpeechRecognition](/docs/api/classes/moonshinespeechrecognition)
-- [StreamTranscriber](/docs/api/classes/streamtranscriber)
+- [Transcriber](/docs/api/classes/transcriber)
 - [VoiceController](/docs/api/classes/voicecontroller)
