+++ src/index.ts
@@ -1,9 +1,8 @@
 import MoonshineModel from "./model";
 import MediaElementTranscriber from "./mediaElementTranscriber"
 import MicrophoneTranscriber from "./microphoneTranscriber";
-import StreamTranscriber from "./streamTranscriber";
 import MoonshineSpeechRecognition from "./webSpeechPolyfill";
-import { TranscriberCallbacks } from "./transcriber";
+import { Transcriber, TranscriberCallbacks } from "./transcriber";
 import { Settings } from "./constants";
 import { VoiceController, KeywordSpotter, IntentClassifier } from "./voiceController";
 import { MoonshineError } from "./error";
@@ -13,9 +12,9 @@ export {
     MoonshineError,
     Settings,
     MoonshineSpeechRecognition,
+    Transcriber,
     MicrophoneTranscriber,
     MediaElementTranscriber,
-    StreamTranscriber,
     TranscriberCallbacks,
     VoiceController,
     KeywordSpotter,
